pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hotbunk"
version = "0.1.0"
description = "Cooperative compute orchestrator for Claude Code"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "httpx>=0.28",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
hotbunk = "hotbunk.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
